Manage Volumes #37

+++ app/Http/Controllers/V1/VolumeController.php
@@ -3,26 +3,15 @@
 namespace App\Http\Controllers\V1;
 
 use App\Http\Controllers\BaseController;
-use App\Http\Requests\AddIDParameterTrait;
-use App\Http\Requests\Ticket\CloseTicketRequest;
-use App\Http\Requests\Ticket\NewReplyRequest;
-use App\Http\Requests\Ticket\NewTicketRequest;
-use App\Http\Requests\Ticket\ShowTicketRequest;
+use App\Http\Requests\Volume\AttachToMachineRequest;
+use App\Http\Requests\Volume\DetachFromMachineRequest;
 use App\Http\Requests\Volume\RemoveVolumeRequest;
 use App\Http\Requests\Volume\RenameVolumeRequest;
-use App\Models\Category;
-use App\Models\Reply;
-use App\Models\Ticket;
-use App\Notifications\NewTicketNotification;
-use App\Notifications\TicketReplyNotification;
-use App\Notifications\TicketStatusNotification;
-use App\Repositories\SSHKeyRepository;
-use App\Repositories\TicketRepository;
+use App\Http\Requests\Volume\CreateVolumeRequest;
+use App\Models\Machine;
 use App\Repositories\VolumeRepository;
 use App\Services\VolumeService;
-use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Str;
 
 class VolumeController extends BaseController
 {
@@ -55,6 +44,144 @@ class VolumeController extends BaseController
     {
         $volumes = $this->repository->all();
         return responder()->success(['list' => $volumes]);
+    }
+
+    /**
+     * @OA\Post(
+     *      tags={"Volume"},
+     *      path="/volumes/createVolume",
+     *      summary="Creates a volume",
+     *      description="",
+     *
+     * @OA\Response(
+     *         response="default",
+     *         description="successful operation"
+     *     ),
+     *
+     *     @OA\Parameter(
+     *         name="name",
+     *         in="query",
+     *         description="",
+     *         required=true,
+     *         @OA\Schema(
+     *             type="string"
+     *         )
+     *     ),
+     *
+     *     @OA\Parameter(
+     *         name="size",
+     *         in="query",
+     *         description="",
+     *         required=true,
+     *         @OA\Schema(
+     *             type="float"
+     *         )
+     *     ),
+     *
+     *     )
+     *
+     */
+    function createVolume(CreateVolumeRequest $request){
+        $service = new VolumeService();
+        $volume = $service->create(\request('name'),\request('size'));
+        $this->repository->create([
+            'remote_id' => $volume->id,
+            'name' => \request('name'),
+            'size' => \request('size'),
+            'is_root' => false,
+            'user_id' => Auth::id()
+        ]);
+        return responder()->success(['message' => 'فضا با موفقیت ساخته شد']);
+    }
+
+    /**
+     * @OA\Post(
+     *      tags={"Volume"},
+     *      path="/volumes/{id}/attachToMachine",
+     *      summary="Attaches a volume to a machine",
+     *      description="",
+     *
+     * @OA\Response(
+     *         response="default",
+     *         description="successful operation"
+     *     ),
+     *
+     *     @OA\Parameter(
+     *         name="id",
+     *         in="path",
+     *         description="",
+     *         required=true,
+     *         @OA\Schema(
+     *             type="int"
+     *         )
+     *     ),
+     *
+     *     @OA\Parameter(
+     *         name="machine_id",
+     *         in="query",
+     *         description="",
+     *         required=true,
+     *         @OA\Schema(
+     *             type="int"
+     *         )
+     *     ),
+     *
+     *     )
+     *
+     */
+    public function attachToMachine(AttachToMachineRequest $request){
+        $volume = $this->repository->find(\request('id'));
+        $machine = Machine::find(\request('machine_id'));
+
+        $service = new VolumeService();
+        $service->attachVolumeToMachine($machine->remote_id, $volume->remote_id);
+
+        return responder()->success(['message' => 'اتصال با موفقیت انجام شد']);
+    }
+
+    /**
+     * @OA\Post(
+     *      tags={"Volume"},
+     *      path="/volumes/{id}/detachFromMachine",
+     *      summary="Detaches a volume from a machine",
+     *      description="",
+     *
+     * @OA\Response(
+     *         response="default",
+     *         description="successful operation"
+     *     ),
+     *
+     *     @OA\Parameter(
+     *         name="id",
+     *         in="path",
+     *         description="",
+     *         required=true,
+     *         @OA\Schema(
+     *             type="int"
+     *         )
+     *     ),
+     *
+     *     @OA\Parameter(
+     *         name="machine_id",
+     *         in="query",
+     *         description="",
+     *         required=true,
+     *         @OA\Schema(
+     *             type="int"
+     *         )
+     *     ),
+     *
+     *     )
+     *
+     */
+    function detachFromMachine(DetachFromMachineRequest $request){
+        $volume = $this->repository->find(\request('id'));
+        $machine = Machine::find(\request('machine_id'));
+
+        $service = new VolumeService();
+        $service->detachVolumeFromMachine($machine->remote_id, $volume->remote_id);
+
+        return responder()->success(['message' => 'قطع ارتباط با موفقیت انجام شد']);
     }
 
     /**
@@ -98,10 +225,12 @@ class VolumeController extends BaseController
         $this->repository->update([
             'name' => \request('name')
         ], \request('id'));
+
+        return responder()->success(['message'=>'نام فضا با موفقیت تغییر یافت']);
     }
 
     /**
-     * @OA\Post(
+     * @OA\Delete(
      *      tags={"Volume"},
      *      path="/volumes/{id}/remove",
      *      summary="Removes the volume",
@@ -128,8 +257,15 @@ class VolumeController extends BaseController
     function remove(RemoveVolumeRequest $request)
     {
         $volume = $this->repository->find(\request('id'));
+
+        if(!empty($volume->machine_id )){
+            return responder()->error(500,'لطفا قبل از حذف فضا اتصال آن را قطع کنید');
+        }
+
         $service = new VolumeService();
         $service->remove($volume->remote_id);
         $this->repository->delete(\request('id'));
+
+        return responder()->success(['message'=>'فضا با موفقیت حذف شد']);
     }
 }
